# api/openapi.yaml
openapi: 3.1.0
info:
  title: Alexandria User Service API
  description: >-
    User account management. Owns the users table; exposes internal fan-out
    endpoints consumed by knowledgebase-service and qa-service.
  license:
    name: MIT
    identifier: MIT
  version: 2.0.0
servers:
  - url: /
    description: Current server
...
paths:
  /api/v1/users/me/preferences:
    patch:
      tags:
        - User Service
      summary: Update user preferences
      operationId: updateUserPreferences
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
        required: true
      responses:
        '200':
          description: User preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferencesDto'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
  /api/v1/users/me:
    get:
      tags:
        - User Service
      summary: Get user profile
      operationId: userProfile
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /api/v1/users/{id}:
    delete:
      tags:
        - User Service
      summary: Delete user account
      description: Users can only delete their own account
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
        '403':
          description: Cannot delete other users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
  /internal/knowledgebase/users/{subject}:
    delete:
      tags:
        - KnowledgeBase Service (Internal)
      summary: Purge all knowledgebase data for a user (internal)
      operationId: kbInternalDeleteUserData
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User data purged
      security: []
...
  /api/v1/qa/ask:
    post:
      tags:
        - QA Service
      summary: Ask a question about your documents
      operationId: ask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskRequest'
        required: true
      responses:
        '200':
          description: Answer generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QAInteraction'
      security:
        - bearerAuth: []
...
  /api/v1/qa/history:
    get:
      tags:
        - QA Service
      summary: Get Q&A history
      operationId: getHistory
      responses:
        '200':
          description: Q&A history retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QAInteraction'
      security:
        - bearerAuth: []
    delete:
      tags:
        - QA Service
      summary: Delete Q&A history
      operationId: deleteHistory
      responses:
        '204':
          description: Q&A history deleted
      security:
        - bearerAuth: []
  /internal/qa/users/{subject}:
    delete:
      tags:
        - QA Service (Internal)
      summary: Purge all QA data for a user (internal)
      operationId: qaInternalDeleteUserData
      parameters:
        - name: subject
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User data purged
      security: []
...
components:
  schemas:
    UpdatePreferencesRequest:
      type: object
      properties:
        darkTheme:
          type: boolean
        language:
          type: string
          maxLength: 5
          minLength: 2
    UserPreferencesDto:
      type: object
      properties:
        darkTheme:
          type: boolean
        language:
          type: string
    UserProfileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        oidcSubject:
          type: string
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        preferences:
          $ref: '#/components/schemas/UserPreferencesDto'
    ErrorResponse:
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: object
...
    AskRequest:
      type: object
      properties:
        question:
          type: string
    QAInteraction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userSubject:
          type: string
        question:
          type: string
        answer:
          type: string
        sourceObjectKeys:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        modelUsed:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...